requirements.txt
alembic==1.9.2
APScheduler==3.10.0
backports.zoneinfo==0.2.1
click==8.1.3
Flask==2.2.2
Flask-APScheduler==1.12.4
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.2
fpdf==1.7.2
greenlet==2.0.2
gunicorn==20.1.0
importlib-metadata==6.0.0
importlib-resources==5.10.2
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
migrate==0.3.8
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SQLAlchemy==2.0.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
Werkzeug==2.2.2
zipp==3.12.0
